Added Answer Box
Added the answer box to ask the player if they want to use another Pokémon after their Pokémon faints in a wild battle. Also fixed the Pokémon menu always saying "CHOOSE A POKéMON" whenever it is opened as well as backing out of this menu when the player is supposed to be forced to send out a new Pokémon.