Multiple additions
Added recurrent damage (burn, poison, and leech) that is applied at the end of a Pokémon's turn. Added checks to the beginning of a Pokémon's attempted move execution for sleep, freeze, confusion, and paralysis, along with their respective effects. Changed the disable check to not reduce the counter if the user is recharging. Added a check after reducing the disable counter to display a message if the user is no longer disabled. Added several messages.