Several changes to Battle Pokémon
Added properties with getters to replace the methods that return whether a Pokémon has a certain status or not. Changed non-volatile status methods to apply the status to both the Battle Pokémon and the Reference Pokémon, as well as to notify the battle manager that this Pokémon's status has changed. Changed the sleep method to require a number of turns the user will sleep for.